requirements.txt
coverage==7.4.0
Jinja2==3.1.3
MarkupSafe==2.1.3
PyYAML==6.0.1
